requirements.txt
crewai==1.9.3
crewai-tools==1.9.3
gdown==5.2.1
langchain==1.2.10
langchain-core==1.2.16
langchain-tavily==0.2.17
pillow==12.1.1
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
streamlit==1.54.0
unstructured==0.18.32
unstructured-client==0.42.10
